requirements.txt
# Core dependencies
datasets>=2.14.0
anthropic>=0.7.0
pyyaml>=6.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
